🐛(utils) correct jaccard similarity measure
The jaccard similarity measure was incorrect.
We used previously the ensemble formula by default:
|X ∩ Y| / (|X| + |Y| - |X ∩ Y|

However, the multicons implementation requires the
pair-wise jaccard similarity measure which is computed
with the following formula:
yy / (yy + ny + yn)

Where:
`yy` is number of times two points belong to same cluster in both
partitions.
`ny` and `yn` are the number of times two points belong to the same
cluster in one partition but not in the other.