✨(consensus) add consensus_function_12
The consensus_function_12 is an enhancement over the regular
consensus_function_10, presented in the thesis of Atheer.
It achieves some times better results than the previous function.
We also fix a bug in the `build_bi_clust` function which should
return the itemsets in sorted order to follow the original algorithm
pseudo-code.